shortify_paths improve 2
