+++ src/diffs.rs
@@ -3,12 +3,12 @@ use std::path::PathBuf;
 use std::sync::Arc;
 use serde::Serialize;
 
-use tokio::sync::{RwLock as ARwLock};
+use tokio::sync::RwLock as ARwLock;
 use hashbrown::{HashMap, HashSet};
 use tracing::info;
-use crate::at_commands::at_file::{file_repair_candidates, get_project_paths, real_file_path_candidate};
+use crate::at_commands::at_file::{file_repair_candidates, real_file_path_candidate};
 use crate::call_validation::DiffChunk;
-use crate::files_correction::correct_to_nearest_dir_path;
+use crate::files_correction::{get_project_paths, correct_to_nearest_dir_path};
 use crate::global_context::GlobalContext;
 
 
@@ -98,7 +98,7 @@ pub async fn correct_and_validate_chunks(
             }
         }
     }
-    
+
     for c in chunks.iter_mut() {
         if c.file_action == "add" {
             c.is_file = PathBuf::from(&c.file_name).extension().is_some() || !c.lines_add.is_empty();
@@ -113,7 +113,7 @@ pub async fn correct_and_validate_chunks(
         }
         validate_chunk(c).map_err(|e| format!("error validating chunk {:?}:\n{}", c, e))?;
     }
-        
+
     Ok(())
 }
 
@@ -237,7 +237,7 @@ fn undo_chunks(
     let mut lines_orig = file_text.split(line_ending).enumerate().map(|(line_n, l)| DiffLine { line_n: line_n + 1, text: l.to_string(), ..Default::default()}).collect::<Vec<_>>();
 
     let mut outputs = HashMap::new();
-    
+
     for (chunk_id, chunk) in chunks.iter().map(|(id, c)|(*id, *c)) {
         let mut chunk_copy = chunk.clone();
 
@@ -321,13 +321,13 @@ pub fn apply_diff_chunks_to_text(
     chunks_undo: Vec<(usize, &DiffChunk)>,
     max_fuzzy_n: usize,
 ) -> (Vec<ApplyDiffResult>, HashMap<usize, ApplyDiffOutput>) {
-    
+
     let mut results = vec![];
     let mut outputs = HashMap::new();
 
     let chunks_apply_edit = chunks_apply.iter().filter(|(_, c)|c.file_action == "edit").cloned().collect::<Vec<_>>();
     let chunks_undo_edit = chunks_undo.iter().filter(|(_, c)|c.file_action == "edit").cloned().collect::<Vec<_>>();
-    
+
     let other_actions = vec!["add", "remove", "rename"];
     let chunks_apply_other = chunks_apply.iter().filter(|(_, c)|other_actions.contains(&c.file_action.as_str())).cloned().collect::<Vec<_>>();
     let chunks_undo_other = chunks_undo.iter().filter(|(_, c)|other_actions.contains(&c.file_action.as_str())).cloned().collect::<Vec<_>>();
@@ -342,14 +342,14 @@ pub fn apply_diff_chunks_to_text(
     ) {
         let line_ending = if file_text.contains("\r\n") { "\r\n" } else { "\n" };
         let mut file_text_copy = file_text.clone();
-        
+
         if chunks_apply_edit.is_empty() && chunks_undo_edit.is_empty() {
             return;
         }
         let file_names = chunks_undo_edit.iter().map(|c|c.1.file_name.clone()).chain(
             chunks_apply_edit.iter().map(|c|c.1.file_name.clone())
         ).collect::<HashSet<_>>().into_iter().collect::<Vec<_>>();
-        
+
         let file_name_edit = match file_names.len() {
             1 => file_names[0].clone(),
             _ => {
@@ -380,7 +380,7 @@ pub fn apply_diff_chunks_to_text(
             ..Default::default()
         });
     }
-    
+
     fn process_chunks_other(
         chunks_apply_other: Vec<(usize, &DiffChunk)>,
         chunks_undo_other: Vec<(usize, &DiffChunk)>,
@@ -392,7 +392,7 @@ pub fn apply_diff_chunks_to_text(
             .into_iter()
             .filter(|c| !undo_ids.contains(&c.0))
             .collect::<Vec<_>>();
-        
+
         if DEBUG == 1 {
             info!("process_chunks_other starts");
             info!("chunks_undo_other_ids: {:?}", chunks_undo_other.iter().map(|c|c.0).collect::<Vec<_>>());
@@ -457,13 +457,13 @@ pub fn apply_diff_chunks_to_text(
                     outputs.insert(c_idx, out);
                 },
                 _ => continue,
-            } 
+            }
         }
     }
 
     process_chunks_edit(chunks_apply_edit, chunks_undo_edit, file_text, max_fuzzy_n, &mut results, &mut outputs);
     process_chunks_other(chunks_apply_other, chunks_undo_other, &mut results, &mut outputs);
-    
+
     (results, outputs)
 }
 
@@ -540,7 +540,7 @@ pub fn read_files_n_apply_diff_chunks(
 }
 
 pub fn unwrap_diff_apply_outputs(
-    outputs: HashMap<usize, ApplyDiffOutput>, 
+    outputs: HashMap<usize, ApplyDiffOutput>,
     chunks_default: Vec<DiffChunk>
 ) -> Vec<ApplyDiffUnwrapped> {
     let mut out_results = vec![];
